minor optimisation for hex encoding
