+++ dan_layer/template_lib_types/src/lib.rs
@@ -5,6 +5,7 @@ pub mod crypto;
 mod entity_id;
 mod error;
 mod hash;
+mod hex;
 pub mod serde_helpers;
 
 pub use entity_id::*;
