+++ dan_layer/template_lib_types/src/serde_helpers.rs
@@ -1,49 +1,13 @@
 //   Copyright 2025 The Tari Project
 //   SPDX-License-Identifier: BSD-3-Clause
 
-use std::{fmt, marker::PhantomData};
-
 use ::serde::{
     de::{Error, Visitor},
     Deserialize,
     Deserializer,
     Serializer,
 };
-use tari_template_abi::rust::borrow::Cow;
-
-use crate::HashParseError;
-pub(crate) fn fixed_bytes_from_hex<const L: usize>(s: &str) -> Result<[u8; L], HashParseError> {
-    if s.len() != L * 2 {
-        return Err(HashParseError);
-    }
-
-    let mut bytes = [0u8; L];
-    for (i, h) in bytes.iter_mut().enumerate() {
-        *h = u8::from_str_radix(&s[2 * i..2 * (i + 1)], 16).map_err(|_| HashParseError)?;
-    }
-    Ok(bytes)
-}
-
-pub(crate) fn bytes_from_hex(s: &str) -> Result<Vec<u8>, HashParseError> {
-    if s.len() % 2 != 0 {
-        return Err(HashParseError);
-    }
-
-    let mut bytes = Vec::with_capacity(s.len() / 2);
-    for i in (0..s.len()).step_by(2) {
-        let byte = u8::from_str_radix(&s[i..i + 2], 16).map_err(|_| HashParseError)?;
-        bytes.push(byte);
-    }
-    Ok(bytes)
-}
-
-pub(crate) fn bytes_to_hex<T: AsRef<[u8]>>(bytes: T) -> String {
-    let mut hex = String::with_capacity(bytes.as_ref().len() * 2);
-    for byte in bytes.as_ref() {
-        hex.push_str(&format!("{byte:02x}"));
-    }
-    hex
-}
+use tari_template_abi::rust::{borrow::Cow, fmt, marker::PhantomData};
 
 #[derive(Default)]
 struct BytesVisitor<'a>(PhantomData<&'a ()>);
@@ -73,6 +37,8 @@ impl<'a> Visitor<'a> for BytesVisitor<'a> {
 
 pub mod fixed_hex {
     use super::*;
+    use crate::hex::{bytes_to_hex, fixed_bytes_from_hex};
+
     pub fn serialize<S: Serializer, T: AsRef<[u8]>>(v: &T, s: S) -> Result<S::Ok, S::Error> {
         if s.is_human_readable() {
             let st = bytes_to_hex(v.as_ref());
@@ -120,8 +86,8 @@ pub mod fixed_hex {
 }
 
 pub mod dynamic_hex {
-
     use super::*;
+    use crate::hex::{bytes_from_hex, bytes_to_hex};
 
     pub fn serialize<S: Serializer, T: AsRef<[u8]>>(v: &T, s: S) -> Result<S::Ok, S::Error> {
         if s.is_human_readable() {
@@ -156,60 +122,6 @@ pub mod dynamic_hex {
             ))
         })
     }
-
-    pub mod option {
-        use super::*;
-
-        pub fn serialize<S: Serializer, T: AsRef<[u8]>>(v: &Option<T>, s: S) -> Result<S::Ok, S::Error> {
-            if s.is_human_readable() {
-                match v {
-                    Some(value) => {
-                        let st = bytes_to_hex(value.as_ref());
-                        s.serialize_some(&st)
-                    },
-                    None => s.serialize_none(),
-                }
-            } else {
-                match v {
-                    Some(value) => s.serialize_some(value.as_ref()),
-                    None => s.serialize_none(),
-                }
-            }
-        }
-
-        pub fn deserialize<'de, D, T>(d: D) -> Result<Option<T>, D::Error>
-        where
-            D: Deserializer<'de>,
-            for<'a> T: TryFrom<&'a [u8]>,
-        {
-            if d.is_human_readable() {
-                let hex = <Option<Cow<'_, str>> as Deserialize>::deserialize(d)?;
-                match hex {
-                    Some(hex_str) => {
-                        let bytes = bytes_from_hex(&hex_str).map_err(Error::custom)?;
-                        T::try_from(&bytes).map(Some).map_err(|_| {
-                            Error::custom(format!(
-                                "Failed to convert bytes to type: {}",
-                                std::any::type_name::<T>()
-                            ))
-                        })
-                    },
-                    None => Ok(None),
-                }
-            } else {
-                let bytes = <Option<Cow<'_, [u8]>>>::deserialize(d)?;
-                match bytes {
-                    Some(b) => T::try_from(&b).map(Some).map_err(|_| {
-                        Error::custom(format!(
-                            "Failed to convert bytes to type: {}",
-                            std::any::type_name::<T>()
-                        ))
-                    }),
-                    None => Ok(None),
-                }
-            }
-        }
-    }
 }
 
 #[cfg(test)]
@@ -224,16 +136,37 @@ mod tests {
         #[serde(with = "super::dynamic_hex")]
         bytes: Vec<u8>,
         pk: RistrettoPublicKeyBytes,
-        #[serde(with = "super::dynamic_hex::option")]
-        opt_bytes: Option<Vec<u8>>,
     }
 
     #[test]
-    fn decode_encode() {
+    fn encode_decode_cbor() {
         let test_case = TestCase {
             bytes: vec![1, 2, 3, 4, 5],
             pk: RistrettoPublicKeyBytes::from([1; 32]),
-            opt_bytes: Some(vec![1, 2, 3, 4, 5]),
+        };
+        let encoded = tari_bor::encode(&test_case).unwrap();
+        let decoded: TestCase = tari_bor::decode(&encoded).unwrap();
+
+        assert_eq!(test_case.bytes, decoded.bytes);
+        assert_eq!(test_case.pk, decoded.pk);
+
+        let cbor = tari_bor::to_value(&test_case).unwrap();
+        let decoded = tari_bor::from_value::<TestCase>(&cbor).unwrap();
+        assert_eq!(test_case.bytes, decoded.bytes);
+        assert_eq!(test_case.pk, decoded.pk);
+        // Check encoded as bytes
+        let bytes = cbor.as_map().unwrap().first().unwrap().1.as_bytes().unwrap();
+        assert_eq!(bytes, &test_case.bytes);
+        // Check encoded as public key
+        let pk = cbor.as_map().unwrap().get(1).unwrap().1.as_bytes().unwrap();
+        assert_eq!(pk, &test_case.pk.as_bytes());
+    }
+
+    #[test]
+    fn decode_encode_json() {
+        let test_case = TestCase {
+            bytes: vec![1, 2, 3, 4, 5],
+            pk: RistrettoPublicKeyBytes::from([1; 32]),
         };
         let json = serde_json::to_string(&test_case).unwrap();
         let decoded: TestCase = serde_json::from_str(&json).unwrap();
@@ -247,7 +180,6 @@ mod tests {
             json["pk"].as_str().expect("string"),
             "0101010101010101010101010101010101010101010101010101010101010101"
         );
-        assert_eq!(json["opt_bytes"].as_str().expect("string"), "0102030405");
 
         let decoded = tari_bor::decode::<TestCase>(&tari_bor::encode(&test_case).unwrap()).unwrap();
         assert_eq!(test_case.bytes, decoded.bytes);
@@ -257,13 +189,5 @@ mod tests {
         let decoded = tari_bor::from_value::<TestCase>(&cbor).unwrap();
         assert_eq!(test_case.bytes, decoded.bytes);
         assert_eq!(test_case.pk, decoded.pk);
-
-        let test_case = TestCase {
-            bytes: vec![1, 2, 3, 4, 5],
-            pk: RistrettoPublicKeyBytes::from([1; 32]),
-            opt_bytes: None,
-        };
-        let json = serde_json::to_value(&test_case).unwrap();
-        assert!(json["opt_bytes"].is_null());
     }
 }
